Allow multiple event objects in one socket data event
In some cases the device socket responded with a data string that
contained multiple objects seperated by an empty line.
If this construct was passed to JSON.stringify it failed with an
unhandled error as the json structure was not valid.
Splitting the response by empty lines and handling each event fixes
this.

In my case I received the following data string after calling `powerOn`
method on the device:
```
{"id":0, "result":["ok"]}
{"method":"props","params":{"power":"off"}}
```